Fixed wrong integer type

+++ src/effects/rainbow_wave.cpp
@@ -22,7 +22,7 @@ void RainbowWave::run(CRGBSet &led_array, uint8_t hue_array[], uint32_t led_num,
     if (effect_setup)
     {
         // Set initial colors that will be shifted through the main sequence
-        for (uint16_t i = 0; i < led_num; i++)
+        for (uint32_t i = 0; i < led_num; i++)
             hue_array[i] = 255 / led_num * i;
         effect_setup = false;
     }
